requirements.txt
Flask
python-dotenv
psycopg2-binary
gunicorn

Flask>=3.0.0
python-dotenv>=1.0.0